+++ src/main/java/org/dx42/needy/parser/GradleDependencyParser.groovy
@@ -20,6 +20,8 @@ import org.dx42.needy.Dependency
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
+import groovy.transform.Canonical
+
 /**
  * DependencyParser for Gradle Groovy-based DSL files
  *
@@ -41,9 +43,9 @@ class GradleDependencyParser implements DependencyParser {
 			throw new IllegalArgumentException("Parameter source was null")
 		}
 
-		GradleDependencyParser_DslEvaluator dslEvaluator = new GradleDependencyParser_DslEvaluator(applicationName, binding, includeFileDependencies)
+        def dslContext = new DslContext(applicationName, binding, includeFileDependencies)
 		
-		GroovyShell shell = createGroovyShell(dslEvaluator, binding)
+		GroovyShell shell = createGroovyShell(dslContext, binding)
 		try {
 			String normalizedSource = STANDARD_GRADLE_API_IMPORTS + source
 			shell.evaluate(normalizedSource)
@@ -53,22 +55,34 @@ class GradleDependencyParser implements DependencyParser {
 			throw new IllegalStateException("An error occurred compiling: [$source]\n${compileError.message}")
 		}
 
-		List<Dependency> dependencies = dslEvaluator.dependencies
+		List<Dependency> dependencies = dslContext.dependencies
 		LOG.info "dependencies = $dependencies"
 		return dependencies
 	}
 
-	private GroovyShell createGroovyShell(GradleDependencyParser_DslEvaluator dslEvaluator, Map<String, Object> binding) {
-		def callDependencies = { Closure closure ->
-			dslEvaluator.evaluate(closure)
-		}
-		Map predefinedMap = [dependencies:callDependencies] + binding
-		Map bindingMap = predefinedMap.withDefault { name -> return DoNothing.INSTANCE }
+	private GroovyShell createGroovyShell(DslContext dslContext, Map<String, Object> binding) {
+        DependenciesDslEvaluator dependenciesDslEvaluator = new DependenciesDslEvaluator(dslContext)
+        BuildscriptDslEvaluator buildscriptDslEvaluator = new BuildscriptDslEvaluator(dslContext)
+         
+		def callDependencies = { Closure closure -> dependenciesDslEvaluator.evaluate(closure) }
+        def callBuildscript = { Closure closure -> buildscriptDslEvaluator.evaluate(closure) }
+        
+        Map bindingMap = [dependencies:callDependencies, buildscript:callBuildscript] + binding
+        bindingMap = bindingMap.withDefault { name -> return DoNothing.INSTANCE }
+        
 		Binding groovyShellBinding = new Binding(bindingMap)
 
 		return new GroovyShell(this.class.classLoader, groovyShellBinding)
 	}
 	
+}
+
+@Canonical
+class DslContext {
+    String applicationName
+    Map<String, Object> binding
+    boolean includeFileDependencies
+    List<Dependency> dependencies = []
 }
 
 class FileDependency {
@@ -78,23 +92,45 @@ class FileDependency {
     def args
 }
 
-class GradleDependencyParser_DslEvaluator {
+class BuildscriptDslEvaluator {
     
-	private static final Logger LOG = LoggerFactory.getLogger(GradleDependencyParser_DslEvaluator)
+    private static final Logger LOG = LoggerFactory.getLogger(BuildscriptDslEvaluator)
+    
+    final DslContext dslContext
+    
+    BuildscriptDslEvaluator(DslContext dslContext) {
+        this.dslContext = dslContext
+    }
+    
+    void dependencies(Closure closure) {
+        DependenciesDslEvaluator dslEvaluator = new DependenciesDslEvaluator(dslContext)
+        dslEvaluator.evaluate(closure)
+    }
+    
+    def methodMissing(String name, args) {
+        LOG.info("methodMissing: name=$name, args=$args")
+    }
+        
+    void evaluate(Closure closure) {
+        closure.delegate = this
+        closure.setResolveStrategy(Closure.DELEGATE_FIRST)
+        closure.call()
+    }
+
+}
+
+class DependenciesDslEvaluator {
+    
+	private static final Logger LOG = LoggerFactory.getLogger(DependenciesDslEvaluator)
 	
-	final List<Dependency> dependencies = []
-	final String applicationName
-	final Map<String, Object> binding
-	final boolean includeFileDependencies
+    final DslContext dslContext
     final List ignoredMethodNames = ['project']
     
-	GradleDependencyParser_DslEvaluator(String applicationName, Map<String, Object> binding, boolean includeFileDependencies) {
-		this.applicationName = applicationName
-		this.binding = binding
-        this.includeFileDependencies = includeFileDependencies
+    DependenciesDslEvaluator(DslContext dslContext) {
+        this.dslContext = dslContext
         
         this.ignoredMethodNames = ['project']
-        if (!includeFileDependencies) {
+        if (!dslContext.includeFileDependencies) {
             this.ignoredMethodNames += ["files", "fileTree"]
         }
 	}
@@ -118,19 +154,19 @@ class GradleDependencyParser_DslEvaluator {
 		}
         if (args[0] instanceof FileDependency) {
             String methodString = args[0].name + "(" + args[0].args.inspect()+ ")"
-            dependencies << new Dependency(applicationName:applicationName, group:FileDependency.GROUP, name:methodString, version:"", configuration:name)
+            dslContext.dependencies << new Dependency(applicationName:dslContext.applicationName, group:FileDependency.GROUP, name:methodString, version:"", configuration:name)
         }
 		if (args[0] instanceof Map) {
 			LOG.info "methodMissing (Map): name=$name value=${args[0]}"
 			for (Map m: args) {
-				dependencies << new Dependency(applicationName:applicationName, group:m.group, name:m.name, version:m.version, configuration:name)
+				dslContext.dependencies << new Dependency(applicationName:dslContext.applicationName, group:m.group, name:m.name, version:m.version, configuration:name)
 			}
 		}
 		else if (args[0] instanceof List) {
 			LOG.info "methodMissing (List): name=$name value=${args[0]}"
 			for (List list: args) {
 				for (String s: list) {
-					dependencies << ParseUtil.createDependencyFromString(applicationName, name, s)
+					dslContext.dependencies << ParseUtil.createDependencyFromString(dslContext.applicationName, name, s)
 				}
 			}
 		}
@@ -139,14 +175,14 @@ class GradleDependencyParser_DslEvaluator {
 			while (index < args.length && args[index] instanceof CharSequence) {
 				String s = args[index]
 				LOG.info "methodMissing: name=$name value=${s}"
-				dependencies << ParseUtil.createDependencyFromString(applicationName, name, s)
+				dslContext.dependencies << ParseUtil.createDependencyFromString(dslContext.applicationName, name, s)
 				index++
 			}
 		}
 	}
 
 	def propertyMissing(String name) {
-		def bindingValue = binding.containsKey(name) ? binding[name] : DoNothing.INSTANCE
+		def bindingValue = dslContext.binding.containsKey(name) ? dslContext.binding[name] : DoNothing.INSTANCE
 		LOG.info("propertyMissing: $name; value=$bindingValue") 
 		return bindingValue
 	}